Add average calculations and various changes
Add averages for subjects, modules, domains.

Add more checks and logic for subjects in order to properly render them.

Separated school and non-school modules.

Add utils for round numbers and calculate averages.

Add a subtext below subjects in annual average in order to explicit what the numbers below the average are.

Add a subtext below domains to tell whether it is eliminatory.